Move filters and functions to separate modules
Currently the index file is rather large and contains a large number of
functions, and this PR instead moves filters and functions to separate
modules, allowing for less noise in the main file and ease of adding new
functions.

Functions exported from functions.js, or filters exported from
filters.js are automatically made available in the templates using the
name of the function (e.g. exporting `my_func` from functions.js will add
a function to the templates called `my_func`.)
