+++ lib/importer/src/filters.js
@@ -1,0 +1,13 @@
+
+
+const currency = (content) => {
+    const num = parseInt(content)
+    if (num === undefined) {
+        return content
+    }
+
+    return "£" + num.toLocaleString()
+}
+
+
+module.exports = { currency }
